Added option to place per-determined RLE patterns
The script `draw_input_matplotlib.py` now includes the ability to read in RLE-formatted patterns and place them within the domain, including a convenient mouse-over feature which shows a transparent outline of the chosen pattern.

Also re-organized directory with an `entries` folder for text files of peoples' tournament entries and a `patterns` directory that contained some RLE-formatted files for common patterns in Conway's Game Of Life.